// libs/api-client/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type RegisterDto = {
    email: string;
    password: string;
    name: string;
};

export type UserResponseDto = {
    id: string;
    email: string;
    name: string;
};

export type AuthResponseDto = {
    accessToken: string;
    refreshToken: string;
    tokenType: string;
    /**
     * Access token lifetime in seconds
     */
    expiresIn: number;
    user: UserResponseDto;
};

export type ErrorResponseDto = {
    statusCode: number;
    message: string;
};

export type LoginDto = {
    email: string;
    password: string;
};

export type RefreshRequestDto = {
    refreshToken: string;
};

export type LogoutRequestDto = {
    refreshToken: string;
};

export type PatientResponseDto = {
    id: string;
    mrn: string;
    firstName: string;
    lastName: string;
    dateOfBirth: string;
    gender: 'MALE' | 'FEMALE' | 'OTHER';
    email: string;
    createdAt: string;
};

export type gender = 'MALE' | 'FEMALE' | 'OTHER';

export type CreatePatientDto = {
    firstName: string;
    lastName: string;
    dateOfBirth: string;
    gender: 'MALE' | 'FEMALE' | 'OTHER';
    email: string;
};

export type UpdatePatientDto = {
    firstName?: string;
    lastName?: string;
    dateOfBirth?: string;
    gender?: 'MALE' | 'FEMALE' | 'OTHER';
    email?: string;
};

export type RegisterData = {
    body: RegisterDto;
};

export type RegisterResponse = (AuthResponseDto);

export type RegisterError = (ErrorResponseDto);

export type LoginData = {
    body: LoginDto;
};

export type LoginResponse = (AuthResponseDto);

export type LoginError = (ErrorResponseDto);

export type RefreshTokenData = {
    body: RefreshRequestDto;
};

export type RefreshTokenResponse = (AuthResponseDto);

export type RefreshTokenError = (ErrorResponseDto);

export type LogoutData = {
    body: LogoutRequestDto;
};

export type LogoutResponse = (void);

export type LogoutError = unknown;

export type ListPatientsResponse = (Array<PatientResponseDto>);

export type ListPatientsError = unknown;

export type CreatePatientData = {
    body: CreatePatientDto;
};

export type CreatePatientResponse = (PatientResponseDto);

export type CreatePatientError = (ErrorResponseDto);

export type GetPatientData = {
    path: {
        id: string;
    };
};

export type GetPatientResponse = (PatientResponseDto);

export type GetPatientError = (ErrorResponseDto);

export type UpdatePatientData = {
    body: UpdatePatientDto;
    path: {
        id: string;
    };
};

export type UpdatePatientResponse = (PatientResponseDto);

export type UpdatePatientError = (ErrorResponseDto);

export type DeletePatientData = {
    path: {
        id: string;
    };
};

export type DeletePatientResponse = (void);

export type DeletePatientError = (ErrorResponseDto);